+++ R/functions-msp.R
@@ -2,9 +2,20 @@
 #'
 #' @description
 #' 
-#' The `readMsp` function imports the data from a file in MGF format reading
+#' The `readMsp()` function imports the data from a file in MGF format reading
 #' all specified fields and returning the data as a [DataFrame()].
 #'
+#' Format constraints for MSP files:
+#'
+#' - Comment lines are expected to start with a `#`.
+#' - Multiple spectra within the same MSP file are separated by an empty line.
+#' - The first n lines of a spectrum entry represent metadata.
+#' - Metadata is provided as "name: value" pairs (i.e. name and value separated
+#'   by a ":").
+#' - One line per mass peak, with values separated by a whitespace or tabulator.
+#' - Each line is expected to contain at least the m/z and intensity values (in
+#'   that order) of a peak. Additional values are currently ignored.
+#' 
 #' @param f `character(1)` with the path to an MSP file.
 #'
 #' @param msLevel `numeric(1)` with the MS level. Default is 2. This value will
@@ -40,28 +51,31 @@
 #'
 #' @examples
 #'
-#' fls <- dir(system.file("extdata", package = "MsBackendMsp"),
-#'     full.names = TRUE, pattern = "msp$")[1L]
+#' f <- system.file("extdata", "minimona.msp", package = "MsBackendMsp")
 #'
-#' readMsp(fls)
+#' readMsp(f)
 readMsp <- function(f, msLevel = 2L,
                     mapping = spectraVariableMapping(MsBackendMsp()), ...) {
     if (length(f) != 1L)
         stop("Please provide a single msp file.")
     
-    msp <- scan(file = f, what = "",
-                sep = "\n", quote = "",
-                allowEscapes = FALSE,
-                quiet = TRUE)
+    msp <- scan(file = f, what = "", sep = "\n", quote = "",
+                allowEscapes = FALSE, quiet = TRUE,
+                blank.lines.skip = FALSE, strip.white = TRUE)
     
     ## Ignore comments
     cmts <- grep("^[#]", msp)
     if (length(cmts))
         msp <- msp[-cmts]
 
-    ## Find individual records
-    begin <- grep("^NAME:", msp, ignore.case = TRUE)
-    end <- c(begin[-1] -1L, length(msp))
+    ## Find individual records. Instead of grepping by NAME: we use blank
+    ## lines. These are expected to separate entries.
+    wsp <- grep("^[[:space:]]|(^$)", msp)
+    begin <- c(1, wsp +1L)
+    end <- c(wsp -1L, length(msp))
+    keep <- begin < end # drop consecutive blank lines.
+    begin <- begin[keep]
+    end <- end[keep]
 
     sp <- mapply(begin, end, FUN = function(a, b) {
          .extract_msp_spectrum(msp[a:b], mapping = mapping)
@@ -74,13 +88,13 @@ readMsp <- function(f, msLevel = 2L,
         if (all(lengths(res[[i]]) == 1))
             res[[i]] <- unlist(res[[i]])
         if (any(col <- names(spv) == colnames(res)[i]))
-            res[[i]] <- as(res[[i]], spv[col][1])
+            res[[i]] <- suppressWarnings(as(res[[i]], spv[col][1]))
     }
 
     res$mz <- NumericList(res$mz, compress = FALSE)
     res$intensity <- NumericList(res$intensity, compress = FALSE)
     res$dataOrigin <- f
-    if (!any(colnames(res) == msLevel))
+    if (!any(colnames(res) == "msLevel"))
         res$msLevel <- as.integer(msLevel)
     res
 }
@@ -94,25 +108,53 @@ readMsp <- function(f, msLevel = 2L,
 #' 
 #' @importFrom stats setNames
 #'
+#' @note
+#'
+#' https://www.nist.gov/system/files/documents/srd/NIST1aVer22Man.pdf
+#' 
 #' @noRd
 .extract_msp_spectrum <- function(msp, mapping) {
     ## grep description
     desc.idx <- grep(":", msp)
     desc <- msp[desc.idx]
-    spec <- msp[-desc.idx]
 
-    ms <- do.call(rbind, strsplit(sub("^(\\t|[[:space:]]+)", "", spec),
-                                  "[[:space:]]+"))
+    l <- length(desc.idx)
+    if (desc.idx[1L] != 1L || desc.idx[l] != l)
+        stop("MSP format error. Make sure that data for multiple spectra are ",
+             "separated by an empty new line and that general spectrum ",
+             "metadata/information is provided in 'name: value' format ",
+             "(i.e. name and value of the metadata field separated by ",
+             "a \":\").", call. = FALSE)
+    
+    spec <- trimws(msp[-desc.idx], "left")
+    
+    pks <- strsplit(sub("^(\\t|[[:space:]]+)", "", spec),
+                    "[[:space:]]+")
+    anns <- lengths(pks) > 2
+    if (any(anns)) {
+        warning("Unexpected number of values per peak found. Keeping only the ",
+                "first two values assuming they correspond to m/z and ",
+                "intensity", call. = FALSE)
+        pks[anns] <- lapply(pks[anns], function(z) z[1:2])
+    }
+    ms <- do.call(rbind, pks)
     mode(ms) <- "double"
-
+    
     if (!length(ms))
         ms <- matrix(numeric(), ncol = 2L)
-    else if (is.unsorted(ms[, 1L]))
-        ms <- ms[order(ms[, 1L]), ]
+    else {
+        if (anyNA(ms[, 1L]))
+            stop("MSP format error. Missing values (NA) for m/z are not ",
+                 "supported. Please ensure that no missing or non-numeric ",
+                 "values are reported as the peaks' m/z values.", call. = FALSE)
+        if (is.unsorted(ms[, 1L]))
+            ms <- ms[order(ms[, 1L]), ]
+    }
     
     r <- regexpr(":", desc, fixed = TRUE)
     desc <- setNames(substring(desc, r + 2L, nchar(desc)),
                      substring(desc, 1L, r - 1L))
+
     ## map fields to spectra variables
     idx <- match(names(desc), mapping)
     not_na <- !is.na(idx)
@@ -189,9 +231,9 @@ readMsp <- function(f, msLevel = 2L,
 #' 
 #' @author Michael Witting, Johannes Rainer
 #'
-#' @importMethodsFrom Spectra spectraVariables spectraNames
+#' @importMethodsFrom Spectra spectraVariables spectraNames spectraData
 #'
-#' @importMethodsFrom Spectra peaksData spectraData
+#' @importMethodsFrom ProtGenerics peaksData
 #'
 #' @noRd
 #'
@@ -255,3 +297,25 @@ readMsp <- function(f, msLevel = 2L,
     tmp[grep(": NA\n", tmp, fixed = TRUE)] <- ""
     writeLines(apply(tmp, 1, paste0, collapse = ""), con = con)
 }
+
+#' @title Parse the comment field from a MoNA MSP file
+#'
+#' @description
+#'
+#' Parse comment field from MoNA.
+#' 
+#' @author Johannes Rainer
+#'
+#' @noRd
+parseMoNaComment <- function(x, names = c("InChI", "author", "SMILES",
+                                          "date", "cas", "kegg",
+                                          "pubchem cid")) {
+    ## extract value between "<name>= and ".
+    names(names) <- names
+    as.data.frame(lapply(names, function(z) {
+        tmp <- sub(paste0(".*?\"", z, "=(.*?)\".*"), "\\1", x, perl = TRUE)
+        tmp[tmp == x] <- NA_character_
+        tmp
+    }))
+}
+